Fixed error in computing block numbers (was int overflow problem).
Added a super instruction check_block_number_calculation and test
that exercises it src/sialx/test/check_block_number_calc.sials to test_sial.

There are two "systems" in the test.  A small one, that is not very useful,
and a large one that reproduced the system where the bug showed
up at scale.  The latter is too big to run as a matter of course and is
commented out.